refactor image decoding
